+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This is just an example changelog used to cover some of the functionalities that
 
 ### To Publish
 
+- **Feature 4:** Visual Feedback when User's Name is not valid in the Login Page
+
 ### Version 1
 
 - **Feature 3:** Breweries Page and State Management